requirements.txt
amqp==2.5.2
appdirs==1.4.3
appnope==0.1.0
apturl==0.5.2
asgiref==3.3.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
billiard==3.6.1.0
bleach==3.1.5
blinker==1.4
Brlapi==0.7.0
celery==4.4.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.1.7
django-common-helpers==0.9.2
django-cors-headers==3.2.1
djangorestframework==3.9.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.3.0
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keyring==18.0.1
kombu==4.6.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
notebook==6.0.3
numpy==1.16.2
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandas==1.2.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
psycopg2==2.8.6
ptyprocess==0.6.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.16.0
pysftp==0.2.9
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.0
python-debian===0.1.36ubuntu1
pytz==2018.9
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
systemd-python==234
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
xkit==0.0.0
xlrd==1.2.0
zipp==1.2.0
